Пиццы сохраняются в local Storage, папка для cart
