+++ src/components/Header.tsx
@@ -4,13 +4,22 @@ import { Link, useLocation } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import Search from './Search';
 
-import { cartSelector } from '../redux/slises/cartSlice';
+import { cartSelector } from '../redux/slises/cart/selectors';
 
 const Header: React.FC = () => {
   const { items, totalPrice } = useSelector(cartSelector);
+  const location = useLocation();
+  const isMounted = React.useRef(false);
+
   const totalCount = items.reduce((sum: number, item: any) => sum + item.count, 0);
 
-  const location = useLocation();
+  React.useEffect(() => {
+    if (isMounted.current) {
+      const json = JSON.stringify(items);
+      localStorage.setItem('cartPizza', json);
+    }
+    isMounted.current = true;
+  }, [items]);
 
   return (
     <div className="container">
