+++ src/redux/slises/cart/selectors.ts
@@ -1,0 +1,4 @@
+import { RootState } from '../../store';
+
+export const cartSelector = (state: RootState) => state.cart;
+export const cartItemSelector = (id: string) => (state: RootState) => state.cart.items.find((obj) => obj.id === id);
