+++ src/utils/calcTotalPrice.ts
@@ -1,0 +1,5 @@
+import { TCartItem } from '../redux/slises/cart/types';
+
+export const calcTotalPrice = (items: TCartItem[]) => {
+  return items.reduce((sum, obj) => obj.price * obj.count + sum, 0);
+};
